LayoutTests/imported/w3c/web-platform-tests/css/css-pseudo/parsing/marker-supported-properties-in-animation-expected.txt


FAIL Animation of font in ::marker assert_in_array: value "italic normal 500 expanded 15px/35px -webkit-standard" not in array ["italic small-caps 500 expanded 15px / 35px Ahem", "italic small-caps 500 expanded 15px/35px Ahem"]
FAIL Animation of font-family in ::marker assert_equals: expected "Ahem" but got "-webkit-standard"
FAIL Animation of font-feature-settings in ::marker assert_equals: expected "\"smcp\"" but got "normal"
FAIL Animation of font-kerning in ::marker assert_equals: expected "none" but got "auto"
PASS Animation of font-size in ::marker
FAIL Animation of font-size-adjust in ::marker assert_true: font-size-adjust doesn't seem to be supported in the computed style expected true got false
PASS Animation of font-stretch in ::marker
PASS Animation of font-style in ::marker
FAIL Animation of font-synthesis in ::marker assert_equals: expected "none" but got "style weight small-caps"
FAIL Animation of font-variant in ::marker assert_equals: expected "small-caps" but got "tabular-nums"
FAIL Animation of font-variant-caps in ::marker assert_equals: expected "small-caps" but got "normal"
FAIL Animation of font-variant-east-asian in ::marker assert_equals: expected "full-width" but got "normal"
FAIL Animation of font-variant-ligatures in ::marker assert_equals: expected "historical-ligatures" but got "normal"
FAIL Animation of font-variant-numeric in ::marker assert_equals: expected "slashed-zero" but got "tabular-nums"
FAIL Animation of font-variant-position in ::marker assert_equals: expected "sub" but got "normal"
PASS Animation of font-weight in ::marker
PASS Animation of line-height in ::marker
FAIL Animation of white-space in ::marker assert_equals: expected "nowrap" but got "pre"
FAIL Animation of color in ::marker assert_equals: expected "rgb(50, 150, 100)" but got "rgb(0, 0, 0)"
FAIL Animation of text-combine-upright in ::marker assert_true: text-combine-upright doesn't seem to be supported in the computed style expected true got false
PASS Animation of unicode-bidi in ::marker
PASS Animation of direction in ::marker
FAIL Animation of content in ::marker assert_equals: expected "\"bar\"" but got ""
FAIL Animation of hyphens in ::marker assert_true: hyphens doesn't seem to be supported in the computed style expected true got false
FAIL Animation of letter-spacing in ::marker assert_equals: expected "10px" but got "normal"
FAIL Animation of line-break in ::marker assert_equals: expected "anywhere" but got "auto"
FAIL Animation of overflow-wrap in ::marker assert_equals: expected "anywhere" but got "normal"
FAIL Animation of tab-size in ::marker assert_equals: expected "10px" but got "8"
FAIL Animation of text-transform in ::marker assert_equals: expected "uppercase" but got "none"
FAIL Animation of word-break in ::marker assert_equals: expected "break-word" but got "normal"
FAIL Animation of word-spacing in ::marker assert_equals: expected "10px" but got "0px"
FAIL Animation of text-decoration-skip-ink in ::marker assert_equals: expected "none" but got "auto"
FAIL Animation of text-emphasis in ::marker assert_equals: expected "triangle rgb(50, 100, 100)" but got ""
FAIL Animation of text-emphasis-color in ::marker assert_equals: expected "rgb(50, 100, 100)" but got "rgb(0, 0, 0)"
FAIL Animation of text-emphasis-position in ::marker assert_equals: expected "under left" but got "over right"
FAIL Animation of text-emphasis-style in ::marker assert_equals: expected "triangle" but got "none"
FAIL Animation of text-shadow in ::marker assert_equals: expected "rgb(50, 100, 100) 2px 2px 2px" but got "none"
PASS Animation of display in ::marker
PASS Animation of position in ::marker
PASS Animation of float in ::marker
FAIL Animation of list-style in ::marker assert_equals: expected "outside none disc" but got "disc outside none"
PASS Animation of list-style-image in ::marker
PASS Animation of list-style-position in ::marker
PASS Animation of list-style-type in ::marker
PASS Transition of font in ::marker
PASS Transition of font-family in ::marker
PASS Transition of font-feature-settings in ::marker
FAIL Transition of font-kerning in ::marker assert_equals: expected "none" but got "auto"
PASS Transition of font-size in ::marker
FAIL Transition of font-size-adjust in ::marker assert_true: font-size-adjust doesn't seem to be supported in the computed style expected true got false
PASS Transition of font-stretch in ::marker
PASS Transition of font-style in ::marker
PASS Transition of font-synthesis in ::marker
PASS Transition of font-variant in ::marker
PASS Transition of font-variant-caps in ::marker
PASS Transition of font-variant-east-asian in ::marker
PASS Transition of font-variant-ligatures in ::marker
PASS Transition of font-variant-numeric in ::marker
PASS Transition of font-variant-position in ::marker
PASS Transition of font-weight in ::marker
PASS Transition of line-height in ::marker
PASS Transition of white-space in ::marker
PASS Transition of color in ::marker
FAIL Transition of text-combine-upright in ::marker assert_true: text-combine-upright doesn't seem to be supported in the computed style expected true got false
PASS Transition of unicode-bidi in ::marker
PASS Transition of direction in ::marker
PASS Transition of content in ::marker
FAIL Transition of hyphens in ::marker assert_true: hyphens doesn't seem to be supported in the computed style expected true got false
PASS Transition of letter-spacing in ::marker
PASS Transition of line-break in ::marker
PASS Transition of overflow-wrap in ::marker
PASS Transition of tab-size in ::marker
PASS Transition of text-transform in ::marker
PASS Transition of word-break in ::marker
PASS Transition of word-spacing in ::marker
PASS Transition of text-decoration-skip-ink in ::marker
FAIL Transition of text-emphasis in ::marker assert_equals: expected "triangle rgb(50, 100, 100)" but got ""
FAIL Transition of text-emphasis-color in ::marker assert_equals: expected "rgb(50, 100, 100)" but got "rgb(100, 0, 200)"
PASS Transition of text-emphasis-position in ::marker
FAIL Transition of text-emphasis-style in ::marker assert_equals: expected "triangle" but got "filled triangle"
PASS Transition of text-shadow in ::marker
PASS Transition of display in ::marker
PASS Transition of position in ::marker
PASS Transition of float in ::marker
FAIL Transition of list-style in ::marker assert_equals: expected "outside none disc" but got "disc outside none"
PASS Transition of list-style-image in ::marker
PASS Transition of list-style-position in ::marker
PASS Transition of list-style-type in ::marker
target
